# openapi.yaml
openapi: 3.0.0
info:
  title: URL Shortener API
  description: An API to shorten URLs, resolve shortened URLs, and manage expiration times.
  version: 1.0.0
servers:
  - url: 'http://localhost/api'
    description: Local Development Server
paths:
...
  /auth/refresh:
    post:
      summary: Refresh JWT token
      operationId: refreshToken
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '401':
          description: Failed to refresh token
...
  /update-expiry-time:
    put:
      summary: Update the expiry time of a shortened URL
      operationId: updateExpiryTime
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                short_url:
                  type: string
                expiry_time:
                  type: integer
                  description: Expiry time in minutes (must be greater than 1)
                  minimum: 1
      responses:
        '200':
          description: Expiry time updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: URL not found
        '400':
          description: Invalid expiry time
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT